+++ examples/gen-docs/gen-docs.cpp
@@ -1,3 +1,4 @@
+#include "arg.h"
 #include "common.h"
 
 #include <fstream>
